+++ doc/changelog.rst
@@ -34,6 +34,7 @@ Localization
 Admin
 ~~~~~
 
+- Show orderability errors in package product management
 - Show stocks in package product management
 - Add link to order line product detail page in order editor
 - Add product line quick add to order creator